services/dashboard/requirements.txt
kafka-python>=2.0.2
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
